Moving the DAGUniverse into the DAGMC header file.
