+++ include/openmc/dagmc.h
@@ -10,6 +10,7 @@ extern "C" const bool DAGMC_ENABLED;
 
 #include "DagMC.hpp"
 
+#include "openmc/cell.h"
 #include "openmc/position.h"
 #include "openmc/xml_interface.h"
 
@@ -19,9 +20,24 @@ namespace model {
   extern std::shared_ptr<moab::DagMC> DAG;
 }
 
-//==============================================================================
-// Non-member functions
-//==============================================================================
+class DAGUniverse : public Universe {
+
+public:
+  explicit DAGUniverse(pugi::xml_node node);
+  explicit DAGUniverse(const std::string& filename, bool auto_ids = false);
+
+  void initialize(); //!< Sets up the DAGMC instance and OpenMC internals
+
+  std::map<int32_t, int32_t> read_uwuw_materials(); //!< Reads UWUW materials and returns an ID map
+
+  // Data Members
+  std::string filename_;
+  std::shared_ptr<moab::DagMC> dagmc_instance_; //! DAGMC Instance for this universe
+  int32_t cell_idx_offset_;
+  int32_t surf_idx_offset_;
+  bool adjust_ids_;
+
+};
 
 void load_dagmc_geometry();
 int32_t create_dagmc_universe(const std::string& filename);
