+++ include/openmc/surface.h
@@ -30,15 +30,6 @@ namespace model {
 } // namespace model
 
 //==============================================================================
-// Constants
-//==============================================================================
-
-enum class GeometryType {
-  CSG,
-  DAG
-};
-
-//==============================================================================
 //! Coordinates for an axis-aligned cube that bounds a geometric object.
 //==============================================================================
